+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Added the built in immutable Python type `frozenset` to the memory-viz immutable, and sequence types
 - Refactored `drawSequence` to use literal types to verify the `type` argument type validation
 - Added id validation for 'draw' input to print a warning when duplicate ids or unresolved id references are detected
+- Added `<title>` and `<desc>` attributes to describe the root `<svg>` element and each `DrawnEntity` within generated SVG diagrams, and added `role="graphics-object"` attribute to `<g>` tags
 
 ### 🐛 Bug fixes